Simplify impl
