+++ serde_arrow/src/internal/schema/from_type/mod.rs
@@ -55,62 +55,62 @@ impl<'de, 'a> serde::de::Deserializer<'de> for TraceAny<'a> {
     }
 
     fn deserialize_bool<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_bool()?;
+        self.0.ensure_primitive(GenericDataType::Bool)?;
         visitor.visit_bool(Default::default())
     }
 
     fn deserialize_i8<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_i8()?;
+        self.0.ensure_primitive(GenericDataType::I8)?;
         visitor.visit_i8(Default::default())
     }
 
     fn deserialize_i16<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_i16()?;
+        self.0.ensure_primitive(GenericDataType::I16)?;
         visitor.visit_i16(Default::default())
     }
 
     fn deserialize_i32<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_i32()?;
+        self.0.ensure_primitive(GenericDataType::I32)?;
         visitor.visit_i32(Default::default())
     }
 
     fn deserialize_i64<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_i64()?;
+        self.0.ensure_primitive(GenericDataType::I64)?;
         visitor.visit_i64(Default::default())
     }
 
     fn deserialize_u8<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_u8()?;
+        self.0.ensure_primitive(GenericDataType::U8)?;
         visitor.visit_u8(Default::default())
     }
 
     fn deserialize_u16<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_u16()?;
+        self.0.ensure_primitive(GenericDataType::U16)?;
         visitor.visit_u16(Default::default())
     }
 
     fn deserialize_u32<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_u32()?;
+        self.0.ensure_primitive(GenericDataType::U32)?;
         visitor.visit_u32(Default::default())
     }
 
     fn deserialize_u64<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_u64()?;
+        self.0.ensure_primitive(GenericDataType::U64)?;
         visitor.visit_u64(Default::default())
     }
 
     fn deserialize_f32<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_f32()?;
+        self.0.ensure_primitive(GenericDataType::F32)?;
         visitor.visit_f32(Default::default())
     }
 
     fn deserialize_f64<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_f64()?;
+        self.0.ensure_primitive(GenericDataType::F64)?;
         visitor.visit_f64(Default::default())
     }
 
     fn deserialize_char<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_u32()?;
+        self.0.ensure_primitive(GenericDataType::U32)?;
         visitor.visit_char(Default::default())
     }
 
@@ -138,7 +138,7 @@ impl<'de, 'a> serde::de::Deserializer<'de> for TraceAny<'a> {
     }
 
     fn deserialize_unit<V: Visitor<'de>>(self, visitor: V) -> Result<V::Value> {
-        self.0.ensure_null()?;
+        self.0.ensure_primitive(GenericDataType::Null)?;
         visitor.visit_unit()
     }
 
@@ -147,7 +147,7 @@ impl<'de, 'a> serde::de::Deserializer<'de> for TraceAny<'a> {
         _name: &'static str,
         visitor: V,
     ) -> Result<V::Value> {
-        self.0.ensure_null()?;
+        self.0.ensure_primitive(GenericDataType::Null)?;
         visitor.visit_unit()
     }
